requirements.txt
python-dotenv
psycopg2-binary
django-ninja
django-cors-headers